refactor(runtime): split context and enforce typed transactions
